+++ src/pyannote/audio/core/model.py
@@ -504,7 +504,6 @@ class Model(lightning.LightningModule):
         revision: str | None = None,
         token: str | bool | None = None,
         cache_dir: Path | str | None = None,
-        skip_dependencies: bool = False,
         **kwargs,
     ) -> Optional["Model"]:
         """Load pretrained model
@@ -534,9 +533,6 @@ class Model(lightning.LightningModule):
         kwargs: optional
             Any extra keyword args needed to init the model.
             Can also be used to override saved hyperparameter values.
-        skip_dependencies : bool, optional
-            If True, skip dependency check. Defaults to False.
-            Use at your own risk, as this may lead to unexpected behavior.
 
         Returns
         -------
@@ -603,12 +599,13 @@ class Model(lightning.LightningModule):
             map_location = default_map_location
 
         # load checkpoint using lightning
-        loaded_checkpoint = pl_load(path_to_model_checkpoint, map_location=map_location)
+        loaded_checkpoint = pl_load(
+            path_to_model_checkpoint, map_location=map_location, weights_only=False
+        )
 
-        if not skip_dependencies:
-            # check that the checkpoint is compatible with the current version
-            versions = loaded_checkpoint["pyannote.audio"]["versions"]
-            check_dependencies({"pyannote.audio": versions["pyannote.audio"]}, "Model")
+        # check that the checkpoint is compatible with the current version
+        versions = loaded_checkpoint["pyannote.audio"]["versions"]
+        check_dependencies({"pyannote.audio": versions["pyannote.audio"]}, "Model")
 
         # obtain model class from the checkpoint
         module_name: str = loaded_checkpoint["pyannote.audio"]["architecture"]["module"]
@@ -625,6 +622,7 @@ class Model(lightning.LightningModule):
                 path_to_model_checkpoint,
                 map_location=map_location,
                 strict=strict,
+                weights_only=False,
                 **kwargs,
             )
         except RuntimeError as e:
@@ -643,6 +641,7 @@ class Model(lightning.LightningModule):
                     path_to_model_checkpoint,
                     map_location=map_location,
                     strict=False,
+                    weights_only=False,
                     **kwargs,
                 )
                 return model