+++ src/pyannote/audio/core/pipeline.py
@@ -21,6 +21,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+from __future__ import annotations
 import os
 import warnings
 from collections import OrderedDict
@@ -51,7 +52,6 @@ def expand_subfolders(
     parent_revision: str | None = None,
     cache_dir: Path | str | None = None,
     token: str | None = None,
-    skip_dependencies: bool = False,
 ) -> None:
     """Expand $model subfolders in config
 
@@ -71,15 +71,11 @@ def expand_subfolders(
         Huggingface token to be used for downloading from Huggingface hub.
     cache_dir: Path or str, optional
         Path to the folder where files downloaded from Huggingface hub are stored.
-    skip_dependencies : bool, optional
-        If True, skip dependency check. Defaults to False.
-        Use at your own risk, as this may lead to unexpected behavior.
     """
 
     if isinstance(config, dict):
         for key, value in config.items():
             if isinstance(value, str) and value.startswith("$model/"):
-
                 subfolder = "/".join(value.split("/")[1:])
 
                 # if subfolder contains '@', split it to get revision
@@ -95,7 +91,6 @@ def expand_subfolders(
                     "subfolder": subfolder,
                     "token": token,
                     "cache_dir": cache_dir,
-                    "skip_dependencies": skip_dependencies,
                 }
             else:
                 expand_subfolders(
@@ -104,13 +99,11 @@ def expand_subfolders(
                     parent_revision=parent_revision,
                     token=token,
                     cache_dir=cache_dir,
-                    skip_dependencies=skip_dependencies,
                 )
 
     elif isinstance(config, list):
         for idx, value in enumerate(config):
             if isinstance(value, str) and value.startswith("$model/"):
-
                 subfolder = "/".join(value.split("/")[1:])
 
                 # if subfolder contains '@', split it to get revision
@@ -126,7 +119,6 @@ def expand_subfolders(
                     "subfolder": subfolder,
                     "token": token,
                     "cache_dir": cache_dir,
-                    "skip_dependencies": skip_dependencies,
                 }
 
             else:
@@ -136,7 +128,6 @@ def expand_subfolders(
                     parent_revision=parent_revision,
                     token=token,
                     cache_dir=cache_dir,
-                    skip_dependencies=skip_dependencies,
                 )
 
 
@@ -144,12 +135,11 @@ class Pipeline(_Pipeline):
     @classmethod
     def from_pretrained(
         cls,
-        checkpoint: str | Path,
+        checkpoint: str | Path | dict,
         revision: str | None = None,
         hparams_file: str | Path | None = None,
         token: str | bool | None = None,
         cache_dir: Path | str | None = None,
-        skip_dependencies: bool = False,
     ) -> Optional["Pipeline"]:
         """Load pretrained pipeline
 
@@ -159,7 +149,8 @@ class Pipeline(_Pipeline):
             Pipeline checkpoint, provided as one of the following:
             * path to a local `config.yaml` pipeline checkpoint
             * path to a local directory containing such a file
-            * identifier of a pipeline on huggingface.co model hub
+            * identifier (str) of a pipeline on huggingface.co model hub
+            * dictionary containing the actual content of a config file
         revision : str, optional
             Revision when loading from the huggingface.co model hub.
         hparams_file: Path or str, optional
@@ -167,30 +158,32 @@ class Pipeline(_Pipeline):
             Token to be used for the download.
         cache_dir: Path or str, optional
             Path to the folder where cached files are stored.
-        skip_dependencies : bool, optional
-            If True, skip dependency check. Defaults to False.
-            Use at your own risk, as this may lead to unexpected behavior.
         """
+
+        # if checkpoint is a dict, assume it is the actual content of
+        # a config file
+        if isinstance(checkpoint, dict):
+            if revision is not None:
+                raise ValueError("Revisions cannot be used with local checkpoints.")
+            model_id = Path.cwd()
+            config = checkpoint
+            otel_origin: str = "local"
 
         # if checkpoint is a directory, look for the pipeline checkpoint
         # inside this directory
-        if os.path.isdir(checkpoint):
+        elif os.path.isdir(checkpoint):
             if revision is not None:
                 raise ValueError("Revisions cannot be used with local checkpoints.")
-
             model_id = Path(checkpoint)
             config_yml = model_id / AssetFileName.Pipeline.value
-            revision = None
             otel_origin: str = "local"
 
         # if checkpoint is a file, assume it is the pipeline checkpoint
         elif os.path.isfile(checkpoint):
             if revision is not None:
                 raise ValueError("Revisions cannot be used with local checkpoints.")
-
             model_id = Path(checkpoint).parent
             config_yml = checkpoint
-            revision = None
             otel_origin: str = "local"
 
         # otherwise, assume that the checkpoint is hosted on HF model hub
@@ -218,8 +211,9 @@ class Pipeline(_Pipeline):
                 else "huggingface"
             )
 
-        with open(config_yml, "r") as fp:
-            config = yaml.load(fp, Loader=yaml.SafeLoader)
+        if not isinstance(checkpoint, dict):
+            with open(config_yml, "r") as fp:
+                config = yaml.load(fp, Loader=yaml.SafeLoader)
 
         # expand $model/{subfolder}-like entries in config
         expand_subfolders(
@@ -228,7 +222,6 @@ class Pipeline(_Pipeline):
             parent_revision=revision,
             token=token,
             cache_dir=cache_dir,
-            skip_dependencies=skip_dependencies,
         )
 
         # before 4.x, pyannote.audio pipeline was using "version" key to
@@ -237,10 +230,9 @@ class Pipeline(_Pipeline):
             config["dependencies"] = {"pyannote.audio": config["version"]}
             del config["version"]
 
-        if not skip_dependencies:
-            # check that dependencies are available (in their required version)
-            dependencies: dict[str, str] = config.get("dependencies", dict())
-            check_dependencies(dependencies, "Pipeline")
+        # check that dependencies are available (in their required version)
+        dependencies: dict[str, str] = config.get("dependencies", dict())
+        check_dependencies(dependencies, "Pipeline")
 
         # initialize pipeline
         pipeline_name = config["pipeline"]["name"]
@@ -415,7 +407,23 @@ class Pipeline(_Pipeline):
         """
         raise NotImplementedError()
 
-    def __call__(self, file: AudioFile, **kwargs):
+    def __call__(self, file: AudioFile, preload: bool = False, **kwargs):
+        """Validate file, (optionally) load it in memory, then process it
+
+        Parameters
+        ----------
+        file : AudioFile
+            File to process
+        preload : bool, optional
+            Whether to preload waveform before applying the pipeline.
+        kwargs : keyword arguments, optional
+            Additional keyword arguments passed to `self.apply(...)`
+
+        Returns
+        -------
+        output : Any
+            Whatever `self.apply(...)` returns
+        """
         fix_reproducibility(getattr(self, "device", torch.device("cpu")))
 
         if not self.instantiated:
@@ -441,8 +449,27 @@ class Pipeline(_Pipeline):
 
         file = Audio.validate_file(file)
 
+        # check if the instance has preprocessors and wrap the file if so
         if hasattr(self, "preprocessors"):
             file = ProtocolFile(file, lazy=self.preprocessors)
+
+        # pre-load the audio in memory if requested
+        if preload:
+            # raise error if `waveform`` is already in memory (or will be via a preprocessor)
+            if (
+                "waveform" in getattr(self, "preprocessors", dict())
+                or "waveform" in file
+            ):
+                raise ValueError(
+                    "Cannot preload audio: `waveform` key is already available or will be via a preprocessor."
+                )
+
+            # load waveform in memory (and keep track of its original sample rate)
+            file["waveform"], file["sample_rate"] = Audio()(file)
+
+            # the above line already took care of channel selection,
+            # therefore we remove the `channel` key from the file
+            file.pop("channel", None)
 
         # send file duration to telemetry as well as
         # requested number of speakers in case of diarization
@@ -450,7 +477,7 @@ class Pipeline(_Pipeline):
 
         return self.apply(file, **kwargs)
 
-    def to(self, device: torch.device):
+    def to(self, device: torch.device) -> Pipeline:
         """Send pipeline to `device`"""
 
         if not isinstance(device, torch.device):
@@ -471,3 +498,14 @@ class Pipeline(_Pipeline):
         self.device = device
 
         return self
+
+    def cuda(self, device: torch.device | int | None = None) -> Pipeline:
+        """Send pipeline to (optionally specified) cuda device"""
+        if device is None:
+            return self.to(torch.device("cuda"))
+        elif isinstance(device, int):
+            return self.to(torch.device("cuda", device))
+        else:
+            if device.type != "cuda":
+                raise ValueError("Expected CUDA device. Use `Pipeline.to(device)` for other devices.")
+            return self.to(device)